backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.2
argcomplete==1.10.3
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.8.2
black==23.7.0
blinker==1.6.2
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
compressed-rtf==1.0.6
decorator==4.4.2
docx2txt==0.8
ebcdic==1.1.1
exceptiongroup==1.1.3
extract-msg==0.28.7
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
frozenlist==1.4.0
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
IMAPClient==2.1.0
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
Mako==1.2.4
MarkupSafe==2.1.3
moviepy==1.0.3
multidict==6.0.4
mypy==1.5.1
mypy-extensions==1.0.0
numpy==1.25.2
olefile==0.46
openai==0.27.8
packaging==23.1
pandas==2.0.3
pathspec==0.11.2
pdfminer.six==20191110
Pillow==10.0.0
platformdirs==3.10.0
pluggy==1.2.0
proglog==0.1.10
psycopg2-binary==2.9.7
pycryptodome==3.18.0
PyJWT==2.8.0
pypdf==3.15.0
PyPDF2==3.0.1
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0.1
reportlab==4.0.4
requests==2.31.0
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SpeechRecognition==3.8.1
SQLAlchemy==2.0.21
textract==1.6.5
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
Werkzeug==2.3.6
wikipedia==1.4.0
xlrd==1.2.0
XlsxWriter==3.1.2
yarl==1.9.2
